Notched plots variant
This script does the same as the master, but instead creates notched box
plots to allow for significance analysis.  Since, for rare samples,
notch limits can be negative and Y axis is log, the box plot Y axis is
set to be constrained to the limits of the data itself.